deepface-auth/requirements.txt
wheel>=0.42.0
setuptools>=68.0.0
opencv-python>=4.8.0
deepface>=0.0.79
numpy<2.0.0
matplotlib>=3.7.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
tensorflow-macos==2.15.0; platform_system == "Darwin"
tensorflow==2.15.0; platform_system != "Darwin"
keras>=2.15.0
torch>=1.9.0
torchvision>=0.10.0
dlib>=19.24.0
pandas>=2.0.0
tf-keras>=2.15.0
networkx>=3.1
python-dotenv>=1.0.0
argparse>=1.4.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=10.0.0
ffmpeg-python==0.2.0
requests>=2.31.0
celery>=5.3.0
psycopg2-binary>=2.9.3
pgvector>=0.1.8
redis>=5.0.0
paho-mqtt>=1.6.1
